update to article and some code fixes

+++ eight_rooty_pieces.cpp
@@ -42,7 +42,8 @@ double my_sqrt_bablyonian(double val) {
  */
 double my_sqrt_newtonraphson(double val) {
 
-  long double x = val / 2;
+  // TODO(PMM) should fail?
+  double x = val / 2;
 
   while (fabs((x * x) - val) > (val / 1E9)) {
     // x * x - val is the function for which we seek the root
